Wave 54: 4099 -> 4132 (+33), bytes 68.10% -> 71.32%
33 functions / 18,304 bytes from the large never-attempted seam
(>=256B, no prior work/ dir). 555 B/fn, ~2.5x wave 53.
8 agents, 3 concurrent slots.

Finding of the wave: literal pools are shared ACROSS function
boundaries. sub_0807974C compiled 100.0% identical over the
original's 952 bytes and still failed, because the candidate was
1004 bytes -- b.n + alignment + 12 pool words. In
asm/code-0806CFC8.s those same 12 words live inside the NEXT
function's address range, behind sub_08079B04's opening
`b _08079B38`. Neither function can ever match alone.
`trymatch.py <fn> --unit` closed the whole 5-function unit at
once (2,668 bytes); three of the five had been stuck at ~99.2%
for several waves and were correct the entire time.
Checked whether this is a wider seam -- it is not, exactly one
multi-function unit had all its unmatched members drafted.

Two already-matched functions retyped from caller evidence, both
re-verified: sub_08075298 a3/a4 u16 -> s16 (signed narrowing
lsls/asrs at the only caller), sub_08076F14 return int -> u8.
Byte-neutral; PROMOTE_MODE zero-extends sub-word parameters
regardless of signedness.

Diagnosed handbacks beat the never-hand-back-an-inherited-draft
rule, n=2: sub_0807B574 (98.6%) and sub_0807AE94 (98.3%) both
closed when the orchestrator named the cause first. Not promoted
to a rule.

Five pre-registered cost predictions refuted across waves 53-54,
including "reading an adjacent batch's verified work makes yours
cheaper" -- the batch built to test it went 2/6, the wave's
worst, next to an untutored 5/5 in the same source file.

Both builds reproduce 14dd0b22c894865867aff89e8116b2dffae25605.

+++ src/decomp/c_08076F14.c
@@ -13,7 +13,13 @@
  * The pool-free if/return pair is the long spelling: `return x*x + y*y <= 255;`
  * would keep the true block as a fallthrough and never materialise the 0. */
 
-int sub_08076F14(s16 x, s16 y)
+/* Wave 54 (W54-F): the RETURN type is `u8`, settled from the only caller.
+ * sub_08076F34 truth-tests the result with `lsls r0, r0, #0x18; cmp r0, #0`,
+ * i.e. it re-narrows to a byte, which an `int` return would not produce. The
+ * change is byte-neutral in this body (the two constants need no narrowing) and
+ * fan-in is 1, so nothing else moves. */
+
+u8 sub_08076F14(s16 x, s16 y)
 {
     if (x * x + y * y <= 0xff)
         return 1;
